V1
A lot of head scratching and swearing.

As far as i can test its working as it should be.

What happens when spectating someone that is in pracmode? IDK havent even looked at that yet.

I have also changed so you cant start a new run in pracmode. IDK why you would want to do this anyways...